[rpc] Hoist `LazyBlockHash` into its own TU
This patch hoists the `LazyBlockHash` abstraction into its translation
unit. In addition the abstraction is moved into the `monad` namespace.
