+++ category/rpc/CMakeLists.txt
@@ -23,8 +23,14 @@ add_library(
   monad_rpc
   OBJECT
   # rpc
-  "monad_executor.cpp" "monad_executor.h"
-  "overrides.cpp" "overrides.h" "overrides.hpp")
+  "lazy_block_hash.cpp" 
+  "lazy_block_hash.hpp"
+  "monad_executor.cpp" 
+  "monad_executor.h"
+  "overrides.cpp" 
+  "overrides.h" 
+  "overrides.hpp"
+  )
 
 target_include_directories(monad_rpc PUBLIC ${CATEGORY_MAIN_DIR})
